third_party/blink/web_tests/svg/dynamic-updates/SVGForeignObjectElement-dom-x-attr-expected.txt

SVG 1.1 dynamic update tests
Test passed

Tests dynamic updates of the 'x' attribute of the SVGForeignObjectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foreignObjectElement.getAttribute('x') is "-100"
PASS foreignObjectElement.getAttribute('x') is "100"
PASS successfullyParsed is true

TEST COMPLETE
